+++ packages/expo-sqlite/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 🎉 New features
 
 - Ship an agent skill for SQLite queries, migrations, storage, and platform configuration in the npm package. ([#49867](https://github.com/expo/expo/pull/49867) by [@Kudo](https://github.com/Kudo))
+- [Android][iOS] Add interruptSync() to cancel running operations on a database connection.
 
 ### 🐛 Bug fixes